refactor wikiContent
