+++ src/pages/Wiki/components/Search.jsx
@@ -3,7 +3,7 @@ import Typography from "@material-ui/core/Typography";
 import TextField from "@material-ui/core/TextField";
 import { Link } from "@material-ui/core";
 
-export default function Search({searchResults, handleChange, loadContent}) {
+export default function Search({ searchResults, handleChange, loadContent }) {
   return (
     <div>
       <div className="wikisearch">
