+++ src/pages/Wiki/wiki.css
@@ -5,10 +5,6 @@
   min-height: calc(100vh - 48px);
 }
 
-.wiki .data {
-  padding: 20px;
-}
-
 .wiki .link {
   font-weight: 900;
 }
@@ -16,14 +12,6 @@
 .wbold {
   font-weight: 900;
   color: #fff;
-}
-
-.datasection {
-  margin: 40px 0px;
-}
-
-.sentence {
-  margin: 25px 0;
 }
 
 .text {
@@ -54,5 +42,4 @@
   text-align: center;
   margin: auto;
   border: solid red;
-  
 }
